Implementação do Spring Security

+++ src/main/java/com/ailton/projeto_multidisciplinar/dtos/UsuarioDTO.java
@@ -1,6 +1,6 @@
 package com.ailton.projeto_multidisciplinar.dtos;
 
-public record UsuarioDTO(Integer id,
+public record UsuarioDTO(Long id,
                          String nome,
                          String telefone,
                          String cpf) {
