+++ src/main/java/com/ailton/projeto_multidisciplinar/infrastructure/repository/ItemPedidoRepository.java
@@ -3,5 +3,5 @@ package com.ailton.projeto_multidisciplinar.infrastructure.repository;
 import com.ailton.projeto_multidisciplinar.infrastructure.entitys.ItemPedido;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItemPedidoRepository extends JpaRepository<ItemPedido, Integer> {
+public interface ItemPedidoRepository extends JpaRepository<ItemPedido, Long> {
 }
