+++ src/main/java/com/ailton/projeto_multidisciplinar/infrastructure/repository/PedidoRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 //No repository adiciona o extends JpaRepository para fazer utilização dos métodos no banco de dados.
-public interface PedidoRepository extends JpaRepository<Pedido, Integer> {
+public interface PedidoRepository extends JpaRepository<Pedido, Long> {
 
     // tipo de retorno
    List<Pedido>findByCanalPedido(TipoCanalAtendimento canalPedido);
